requirements.txt
yt-dlp
Pillow
mutagen
requests
PyQt6